Add MachineListItem control
This change adds the machinelist item control which is
a card that shows relevant information with some
useful buttons at the bottom. Eventually, clicking the
card should result in it flipping over and revealing more
info as a popup. It will appear that this info is on the
"back" of the card.